*! Introduce gd::wstring and gd:wchar and switch to them from std::wstring and    wchar_t. This changes nothing on Linux and most other systems, but on Win32    it causes to use normal UCS-4 strings instead of Win32's usual UTF-16.

+++ src/dictionary.cc
@@ -83,7 +83,7 @@ long DataRequest::dataSize()
 {
   Mutex::Lock _( dataMutex );
   
-  return hasAnyData ? data.size() : -1;
+  return hasAnyData ? (long)data.size() : -1;
 }
 
 void DataRequest::getDataSlice( size_t offset, size_t size, void * buffer )
